Update _titlepage.tex
The ANU Thesis Format Policy requires that the title page of the thesis includes the title, the candidate’s full name, and the month and year of submission. It must also contain the statement:
"A thesis submitted for the degree of Doctor of Philosophy of The Australian National University."

In addition, the title page must include a copyright notice centered at the bottom, providing the full legal name of the author and the year of submission:
"© Copyright by [Candidate’s Full Name] [Year]
All Rights Reserved"

https://policies.anu.edu.au/ppl/document/ANUP_012815 

To ensure compliance with these requirements, the ANU Quarto template on GitHub has been updated to reflect the correct title page format.